pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "searchbench"
version = "0.1.0"
description = "Honest benchmarks for agentic search APIs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "httpx==0.27.2",
  "openai==1.57.0",
  "python-dotenv==1.0.1",
  "typer==0.12.3",
  "tomli==2.0.1; python_version < '3.11'",
]

[project.scripts]
searchbench = "searchbench.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["searchbench*"]

[tool.setuptools.package-data]
searchbench = ["queries/*.json", "queries/*.template"]
